\name{Tampo-package}
\alias{Tampo-package}
\alias{Tampo}
\docType{package}
\title{
Tools for the Analysis of Microchemical Profiles from Otoliths
}
\description{
Facilitates analyzing microchemical profiles (both mono- and multielemental composition) from fish otoliths (stones from the inner ear) using a recursive partitioning approach that can accommodate some form of user-specified constraints. It also allows extracting environmental histories from otolith based on typical elemental sequences generated using methods such as LA-ICPMS. Microchemical analysis is widely used in fisheries management and fisheries biology to identify stocks and characterize fish movements but the provided functions may be considered from the more general perspective of the chronological clustering of multivariate time series using piecewise constant regressions.
}
\details{
\tabular{ll}{
Package: \tab Tampo\cr
Type: \tab Package\cr
Version: \tab 1.0\cr
Date: \tab 2014-05-30\cr
License: \tab GPL (>= 2)\cr
}
}
\author{
Matthias Vignon

Maintainer: Matthias Vignon <Matthias.Vignon@univ-pau.fr>

}
\references{
~~ Literature or other references for background information ~~
}
\keyword{ package }
\seealso{
\code{\link[mvpart]{mvpart}}
}
